Add Rescan Lyrics feature and UI enhancements
- Introduced a "Rescan Lyrics" button in `AppShell.xaml` and implemented its click handler in `AppShell.xaml.cs`.
- Updated `Dimmer.Droid.csproj` to remove an unused XAML file reference.
- Modified `MainActivity.cs` to support Picture-in-Picture mode and adjusted foreground service handling.
- Enhanced `BaseViewModelAnd.cs` to allow nullable progress reporter in `LoadSongDataAsync`.
- Simplified layout in `BtmSheetSearchFilterAndSongsColViewUI.xaml` by removing unnecessary XAML code.
- Added `MainViewExpander` and `QuickPanelBtmSheet` components in `HomePage.xaml` for better song selection handling.
- Updated `SettingsPage.xaml` to include the new "Rescan Lyrics" button.
- Improved layout and data binding in `SingleSongPage.xaml`.
- Refactored `BaseViewModel.cs` for better song data loading and event management.
- Enhanced event handling for song selection and UI interactions across various files.
- Cleaned up code by removing commented-out sections and unnecessary complexity.
