+++ Dimmer/Dimmer.Droid/MainActivity.cs
@@ -21,7 +21,7 @@ namespace Dimmer;
 [IntentFilter(new[] { "android.intent.action.MUSIC_PLAYER" },
               Categories = new[] { Intent.CategoryDefault, "android.intent.category.APP_MUSIC" })]
 [Activity(Theme = "@style/Maui.SplashTheme",
-    MainLauncher = true,
+    MainLauncher = true,SupportsPictureInPicture =true,
         Name = "com.yvanbrunel.dimmer.MainActivity",
     LaunchMode = LaunchMode.SingleTask,
     ConfigurationChanges = ConfigChanges.ScreenSize |
@@ -127,9 +127,8 @@ public class MainActivity : MauiAppCompatActivity
         _serviceIntent = new Intent(this, typeof(ExoPlayerService));
         if (Build.VERSION.SdkInt >= BuildVersionCodes.O)
         {
-
-            this.
-            StartForegroundService(_serviceIntent);
+            
+            this.StartForegroundService(_serviceIntent);
 
         }
         else
@@ -155,8 +154,6 @@ public class MainActivity : MauiAppCompatActivity
         Window.SetStatusBarColor(Android.Graphics.Color.ParseColor("#861B2D"));
         Window.SetStatusBarColor(Android.Graphics.Color.ParseColor("#861B2D"));
         Window.SetUiOptions(UiOptions.SplitActionBarWhenNarrow); // Split action bar for narrow screens
-        var ss = Resource.Drawable.appicon;
-        Window.SetIcon(ss); // Set the app icon
         //Window.SetStatusBarColor(Android.Graphics.Color.Transparent); // Make status bar transparent
         // Tells the Window to draw under the status bar
